Changes BigDecimal returns to Number returns

+++ samples/client/petstore/java/src/main/java/org/openapijsonschematools/schemas/DoubleSchema.java
@@ -3,13 +3,12 @@ package org.openapijsonschematools.schemas;
 import org.openapijsonschematools.configurations.SchemaConfiguration;
 
 import java.util.LinkedHashSet;
-import java.math.BigDecimal;
 
 public record DoubleSchema(LinkedHashSet<Class<?>> type, String format) implements Schema {
     public static DoubleSchema withDefaults() {
         LinkedHashSet<Class<?>> type = new LinkedHashSet<>();
-        type.add(BigDecimal.class);
-        String format = "float";
+        type.add(Double.class);
+        String format = "double";
         return new DoubleSchema(type, format);
     }
